+++ NEWS.md
@@ -6,6 +6,7 @@
 - fix `calendarPlot` issue when there are gaps between selected months e.g.`month = c(1, 2, 12)`
 - fix warning message in `calendarPlot`
 - don't pad missing dates in `timeAverage` if DST
+- deprecate `resolution` in `polarPlot` --- now uses predictions at a coarser resolution that are interpolated. Will also speed-up plotting
 
 # openair 2.9-1